requirements.txt
aiogram==3.17.0
pydantic==2.10.6
openai==1.61.0
python-dotenv==1.0.1
tenacity==9.0.0
duckduckgo-search>=7.0.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
requests>=2.31.0
chromadb>=0.4.0
apscheduler>=3.10.0
